requirements.txt
# ===== Data Science & Math =====
# Used globally across local and cloud environments for core logic.
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
statsmodels>=0.14.0

# ===== Machine Learning =====
# Core offline and real-time training engines.
scikit-learn>=1.4.0
lightgbm>=4.3.0
torch>=2.2.0

# ===== Natural Language Processing =====
# Used primarily offline during dataset building to map restaurant text. 
# Requires an active internet connection when running `python scripts/run_pipeline.py`.
sentence-transformers>=2.5.0
google-genai>=0.1.0

# ===== Tuning & Optimization =====
# Used offline for Bayesian hyperparameter searching.
optuna>=4.0.0

# ===== Visualization =====
# Used offline by `evaluate_model.py` to generate .png reports in `reports/`.
matplotlib>=3.8.0
seaborn>=0.13.0

# ===== Web & Server Integration =====
# Cloud inference stack. Required when running `uvicorn src.inference_service:app`.
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
requests>=2.31.0

# ===== Environment Handling =====
# Extremely critical for offline local testing to prevent leaking API keys.
# Required by `inference_service` to load the `GEMINI_API_KEY` from your local `.env` file.
# Note: On Hugging Face Spaces, secrets are injected natively into the Docker environment.
python-dotenv>=1.0.1
